Share new context crease tracking code with the inline prompt editor
